No need to install mingw if it's already installed.

+++ scripts/ci-build.sh
@@ -23,7 +23,9 @@ NDS_PATH=$PWD
 
 if [ "$TARG_ARCH" == "x86_64-w64-mingw32" ]; then
     # see https://github.com/randombit/botan/issues/2039
-    apt-get install -y g++-mingw-w64-x86-64
+    if [ ! -f /usr/bin/x86_64-w64-mingw32-g++-posix ]; then
+        apt-get install -y g++-mingw-w64-x86-64
+    fi
     update-alternatives --set x86_64-w64-mingw32-g++ /usr/bin/x86_64-w64-mingw32-g++-posix
     # CROSS_COMPILE used by dlfcn-win32/tools/ci-build.sh
     export CROSS_COMPILE=$TARG_ARCH
@@ -40,7 +42,9 @@ if [ "$TARG_ARCH" == "x86_64-w64-mingw32" ]; then
         popd
     fi
 elif [ "$TARG_ARCH" == "i686-w64-mingw32" ]; then
-    apt-get install -y g++-mingw-w64-i686
+    if [ ! -f /usr/bin/i686-w64-mingw32-g++-posix ]; then
+        apt-get install -y g++-mingw-w64-i686
+    fi
     # see https://github.com/randombit/botan/issues/2039
     update-alternatives --set i686-w64-mingw32-g++ /usr/bin/i686-w64-mingw32-g++-posix
     export CROSS_COMPILE=$TARG_ARCH
